+++ catalyst_voices/apps/voices/lib/pages/proposals/widgets/proposals_pagination.dart
@@ -1,7 +1,7 @@
 import 'package:catalyst_voices/pages/proposals/widgets/proposals_pagination_empty_state.dart';
 import 'package:catalyst_voices/pages/proposals/widgets/proposals_pagination_tile.dart';
 import 'package:catalyst_voices/widgets/pagination/builders/paged_wrap_child_builder.dart';
-import 'package:catalyst_voices/widgets/pagination/layouts/paginated_grid_view.dart.dart';
+import 'package:catalyst_voices/widgets/pagination/layouts/paginated_grid_view.dart';
 import 'package:catalyst_voices/widgets/pagination/paging_controller.dart';
 import 'package:catalyst_voices_view_models/catalyst_voices_view_models.dart';
 import 'package:flutter/material.dart';
